+++ elephant/__init__.py
@@ -19,6 +19,7 @@ from . import (statistics,
                current_source_density,
                change_point_detection,
                phase_analysis,
+               parallel,
                sta,
                conversion,
                neo_tools,
