+++ elephant/test/test_parallel.py
@@ -5,8 +5,7 @@ import numpy as np
 import quantities as pq
 from numpy.testing import assert_array_almost_equal
 
-from elephant.parallel import SingleProcess, MPIPoolExecutor, \
-    ProcessPoolExecutor, MPICommExecutor
+from elephant.parallel import SingleProcess, ProcessPoolExecutor
 from elephant.spike_train_generation import homogeneous_poisson_process
 from elephant.statistics import mean_firing_rate
 
@@ -16,10 +15,13 @@ python_version_major = sys.version_info.major
 class TestParallel(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        cls.executors_cls = (
-            SingleProcess, ProcessPoolExecutor, MPIPoolExecutor,
-            MPICommExecutor
-        )
+        cls.executors_cls = [SingleProcess, ProcessPoolExecutor]
+        try:
+            from elephant.parallel.mpi import MPIPoolExecutor, MPICommExecutor
+            cls.executors_cls.extend([MPIPoolExecutor, MPICommExecutor])
+        except ImportError:
+            # mpi4py is not installed
+            pass
 
         np.random.seed(28)
         n_spiketrains = 10
@@ -30,6 +32,7 @@ class TestParallel(unittest.TestCase):
         )
         cls.mean_fr = tuple(map(mean_firing_rate, cls.spiketrains))
 
+    @unittest.skipUnless(python_version_major == 3, "subTest requires 3.4")
     def test_mean_firing_rate(self):
         for executor_cls in self.executors_cls:
             with self.subTest(executor_cls=executor_cls):
